+++ packages/expo-build-properties/src/__tests__/pluginConfig-test.ts
@@ -40,7 +40,7 @@ describe(validateConfig, () => {
     expect(() =>
       validateConfig({ ios: { deploymentTarget: '9.0' } })
     ).toThrowErrorMatchingInlineSnapshot(
-      `"\`ios.deploymentTarget\` needs to be at least version 12.0."`
+      `"\`ios.deploymentTarget\` needs to be at least version 13.0."`
     );
   });
 });